phase_2/requirements.txt
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
nltk
Pillow
tqdm